refactor(api): move GET /api/users/my-account to deprecated
