+++ api/src/deprecated/application/users-me.route.js
@@ -13,4 +13,16 @@ export const usersMeRoute = [
       tags: ['api', 'deprecated', 'users-me'],
     },
   },
+  {
+    method: 'GET',
+    path: '/api/users/my-account',
+    config: {
+      handler: (request, h) => usersMeController.getCurrentUserAccountInfo(request, h),
+      notes: [
+        '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+          '- Récupération des informations du compte utilisateur authentifié\n',
+      ],
+      tags: ['api', 'deprecated', 'user', 'my-account'],
+    },
+  },
 ];
