+++ api/src/deprecated/domain/usecases/index.js
@@ -1,3 +1,4 @@
+import * as authenticationMethodRepository from '../../../../src/identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import { legalDocumentApiRepository } from '../../../../src/identity-access-management/infrastructure/repositories/legal-document-api.repository.js';
 import * as userRepository from '../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
 import * as membershipRepository from '../../../../src/team/infrastructure/repositories/membership.repository.js';
@@ -9,13 +10,16 @@ import { userOrgaSettingsRepository } from '../../../team/infrastructure/reposit
 import boundedContext from '../../dependencies.json' with { type: 'json' };
 import * as certificationPointOfContactRepository from '../../infrastructure/repositories/certification-point-of-contact-repository.js';
 import { prescriberRepository } from '../../infrastructure/repositories/prescriber-repository.js';
+import * as privacyUsersApiRepository from '../../infrastructure/repositories/privacy-users-api.repository.js';
 import * as userAdminRepository from '../../infrastructure/repositories/user-admin-repository.js';
 import { getCertificationPointOfContact } from './get-certification-point-of-contact.js';
 import { getCurrentUser } from './get-current-user.js';
 import { getPrescriber } from './get-prescriber.js';
+import { getUserAccountInfo } from './get-user-account-info.usecase.js';
 import { getUserDetailsForAdmin } from './get-user-details-for-admin.usecase.js';
 
 const dependencies = {
+  authenticationMethodRepository,
   userRepository,
   userAdminRepository,
   prescriberRepository,
@@ -25,6 +29,7 @@ const dependencies = {
   userRecommendedTrainingRepository,
   legalDocumentApiRepository,
   certificationPointOfContactRepository,
+  privacyUsersApiRepository,
   centerRepository,
 };
 
@@ -33,6 +38,7 @@ const usecasesWithoutInjectedDependencies = {
   getCertificationPointOfContact,
   getPrescriber,
   getUserDetailsForAdmin,
+  getUserAccountInfo,
 };
 
 export const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies, boundedContext);
