+++ api/src/identity-access-management/application/user/user.route.js
@@ -90,18 +90,6 @@ export const userRoutes = [
   },
   {
     method: 'GET',
-    path: '/api/users/my-account',
-    config: {
-      handler: (request, h) => userController.getCurrentUserAccountInfo(request, h),
-      notes: [
-        '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-          '- Récupération des informations du compte utilisateur authentifié\n',
-      ],
-      tags: ['identity-access-management', 'api', 'user', 'my-account'],
-    },
-  },
-  {
-    method: 'GET',
     path: '/api/users/{id}/authentication-methods',
     config: {
       validate: {
