+++ api/tests/deprecated/integration/domain/usecases/get-user-account-info.usecase.test.js
@@ -1,9 +1,9 @@
-import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
+import { usecases } from '../../../../../src/deprecated/domain/usecases/index.js';
 import { featureToggles } from '../../../../../src/shared/infrastructure/feature-toggles/index.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
 
-describe('Integration | Identity Access Management  | Domain | Usecases | get-user-account-info', function () {
+describe('Integration | Deprecated  | Domain | Usecases | get-user-account-info', function () {
   beforeEach(async function () {
     await featureToggles.set('isSelfAccountDeletionEnabled', false);
   });
